backend/core/requirements.txt
flask
flask-jwt-extended
flask-cors
mysql-connector-python
bcrypt
gunicorn
python-dotenv